Make general functions for Homewizard
